refactoring and reorganizing
