+++ service/src/main/kotlin/fi/espoo/evaka/invoicing/data/FeeDecisionQueries.kt
@@ -33,7 +33,10 @@ import fi.espoo.evaka.shared.utils.splitSearchText
 import java.time.LocalDate
 import java.util.UUID
 
-fun feeDecisionQuery(predicate: Predicate<Any> = Predicate.alwaysTrue()) =
+fun feeDecisionQuery(
+    predicate: Predicate<Any> = Predicate.alwaysTrue(),
+    lockForUpdate: Boolean = false
+) =
     QuerySql.of<Any> {
         sql(
             """
@@ -86,6 +89,7 @@ SELECT
     ), '[]'::jsonb) AS children
 FROM fee_decision as decision
 WHERE ${predicate(predicate.forTable("decision"))}
+${if (lockForUpdate) "FOR UPDATE" else ""}
 """
         )
     }
@@ -641,8 +645,9 @@ fun Database.Read.getFeeDecision(uuid: FeeDecisionId): FeeDecisionDetailed? {
 
 fun Database.Read.findFeeDecisionsForHeadOfFamily(
     headOfFamilyId: PersonId,
-    period: DateRange?,
-    status: List<FeeDecisionStatus>?
+    period: DateRange? = null,
+    status: List<FeeDecisionStatus>? = null,
+    lockForUpdate: Boolean = false
 ): List<FeeDecision> {
     val headPredicate = Predicate<Any> { where("$it.head_of_family_id = ${bind(headOfFamilyId)}") }
     val validPredicate =
@@ -657,7 +662,7 @@ fun Database.Read.findFeeDecisionsForHeadOfFamily(
                 )
             }
     val predicate = Predicate.all(listOf(headPredicate, validPredicate, statusPredicate))
-    return createQuery(feeDecisionQuery(predicate)).toList<FeeDecision>()
+    return createQuery(feeDecisionQuery(predicate, lockForUpdate)).toList<FeeDecision>()
 }
 
 fun Database.Transaction.approveFeeDecisionDraftsForSending(
