+++ service/src/test/kotlin/fi/espoo/evaka/invoicing/service/generator/FeeDecisionGeneratorV2Test.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package fi.espoo.evaka.invoicing.service
+package fi.espoo.evaka.invoicing.service.generator
 
 import fi.espoo.evaka.daycare.domain.ProviderType
 import fi.espoo.evaka.invoicing.createFeeDecisionChildFixture
